+++ src/Credfeto.Dispatcher.GitHub/Models/GraphQlCommitDetail.cs
@@ -1,0 +1,7 @@
+using System.Diagnostics;
+using System.Text.Json.Serialization;
+
+namespace Credfeto.Dispatcher.GitHub.Models;
+
+[DebuggerDisplay("Commit")]
+internal sealed record GraphQlCommitDetail([property: JsonPropertyName("author")] GraphQlCommitAuthor? Author);